Moved function tests to separate file
